+++ tests/examples/path_filter/main.sh
@@ -1,0 +1,39 @@
+#!/usr/bin/env -S bash -x
+# Exit on first error and clean up.
+set -e
+trap 'kill $(pgrep -g $$ | grep -v $$) > /dev/null 2> /dev/null || :' EXIT
+rm -rvf $(cat .gitignore)
+
+# Run the example a first time, leaving the filter to its default value.
+stepup boot -n 1 -w & # > current_stdout1.txt &
+
+# Get the graph after completion of the pending steps.
+stepup wait
+stepup graph current_graph1
+stepup join
+
+# Wait for background processes, if any.
+wait
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1
+grep -v foo paths.txt
+grep bar paths.txt
+
+
+# Run the example a second time, setting the filter to an empty string, which will allow venv.
+export STEPUP_PATH_FILTER=""
+stepup boot -n 1 -w & # > current_stdout2.txt &
+
+# Get the graph after completion of the pending steps.
+stepup wait
+stepup graph current_graph2
+stepup join
+
+# Wait for background processes, if any.
+wait
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1
+grep foo paths.txt
+grep bar paths.txt